Restructure the client API based on https://github.com/ipfs/interface-ipfs-core/tree/master/SPEC
